# tests/unit/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/api/tasks/{task_id}": {
      "get": {
        "tags": ["API · ✅ Задачи"],
        "summary": "Получить задачу по ID",
        "description": "**Получение задачи по ID**\n\n---\n\n- Этот эндпоинт позволяет получить полную информацию о задаче по её ID.\n- Может использоваться для просмотра и проверки деталей отдельной задачи.\n- В случае успешного выполнения возвращаются данные задачи, соответствующей указанному ID.",
        "operationId": "get_task_api_tasks__task_id__get",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "ID задачи",
              "examples": [0, 1, 2]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResponseTasks" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["API · ✅ Задачи"],
        "summary": "Удалить задачу",
        "description": "**Удаление задачи**\n\n---\n\n- Этот эндпоинт позволяет удалить задачу по её ID.\n- После успешного удаления задача полностью удаляется из системы.\n- Операция удаления необратима. После удаления восстановить задачу невозможно.",
        "operationId": "delete_task_api_tasks__task_id__delete",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Task Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/tasks/": {
      "get": {
        "tags": ["API · ✅ Задачи"],
        "summary": "Получить все задачи",
        "description": "**Получение всех задач**\n\n---\n\n- Этот эндпоинт позволяет получить список всех задач пользователя.\n- Поддерживает фильтрацию по статусу выполнения, а также сортировку по дате создания.\n- В случае успешного выполнения возвращается массив всех задач пользователя с полной информацией о каждой задачи.",
        "operationId": "get_all_task_api_tasks__get",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "completed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "description": "Фильтровать задачи по статусу: выполнено или нет",
              "title": "Completed"
            },
            "description": "Фильтровать задачи по статусу: выполнено или нет"
          },
          {
            "name": "order_by_created",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "description": "Сортировать задачи по дате создания",
              "title": "Order By Created"
            },
            "description": "Сортировать задачи по дате создания"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ResponseTasks" },
                  "title": "Response Get All Task Api Tasks  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/tasks/task/{task_id}": {
      "patch": {
        "tags": ["API · ✅ Задачи"],
        "summary": "Изменить задачу",
        "description": "**Обновление существующей задачи**\n\n---\n\n- Этот эндпоинт позволяет обновить информацию о задаче по её ID.\n- Все поля можно изменять\n- Передавайте только те поля, которые необходимо обновить.\n- В случае успешного выполнения возвращается задача с обновлёнными данными.",
        "operationId": "edit_task_api_tasks_task__task_id__patch",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Task Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RequestTask" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResponseTasks" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/tasks/task/{task_id}/old": {
      "patch": {
        "tags": ["API · ✅ Задачи"],
        "summary": "Этот эндпоинт устарел",
        "operationId": "edit_task_old_api_tasks_task__task_id__old_patch",
        "deprecated": true,
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Task Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RequestTask" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResponseTasks" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/tasks/{task_id}": {
      "get": {
        "tags": ["WEB · ✅ Задачи"],
        "summary": "Получить задачу по ID",
        "description": "**Получение задачи по её ID.**\n\n**Параметры запроса:**\n- `task_id`: Уникальный идентификатор задачи (path parameter)\n\n**Требования:**\n- Пользователь должен быть аутентифицирован\n\n**Возвращает:**\n- Созданную задачу со всеми полями\n\n**Возможные ошибки:**\n- **401**: Пользователь не аутентифицирован\n- **404**: Задача не найдена\n- **422**: Невалидный формат task_id\n- **429**: Превышен лимит запросов\n- **500**: Внутренняя ошибка сервера",
        "operationId": "get_task_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "ID задачи",
              "examples": [0, 1, 2]
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResponseTasks" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["WEB · ✅ Задачи"],
        "summary": "Изменить задачу",
        "operationId": "edit_task_tasks__task_id__patch",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "ID задачи",
              "examples": [0, 1, 2]
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RequestTask" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Задача изменена",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResponseTasks" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["WEB · ✅ Задачи"],
        "summary": "Удалить задачу",
        "operationId": "delete_task_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "ID задачи",
              "examples": [0, 1, 2]
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Задача удалена",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/tasks/": {
      "get": {
        "tags": ["WEB · ✅ Задачи"],
        "summary": "Получить все задачи",
        "operationId": "get_all_task_tasks__get",
        "parameters": [
          {
            "name": "completed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "description": "Фильтровать задачи по статусу: выполнено или нет",
              "title": "Completed"
            },
            "description": "Фильтровать задачи по статусу: выполнено или нет"
          },
          {
            "name": "order_by_created",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "description": "Сортировать задачи по дате создания",
              "title": "Order By Created"
            },
            "description": "Сортировать задачи по дате создания"
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Все задачи отправлены",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ResponseTasks" },
                  "title": "Response Get All Task Tasks  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RequestTask": {
        "properties": {
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "description": "Статус выполнения",
            "default": false,
            "examples": [true, false]
          },
          "priority": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": -1.0,
            "title": "Priority",
            "description": "Приоритет задачи (0–3)",
            "examples": [2]
          },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 100, "minLength": 0 },
              { "type": "null" }
            ],
            "title": "Description",
            "description": "Описание задачи",
            "examples": ["Купить молоко, хлеб и яйца в магазине"]
          },
          "title": {
            "anyOf": [
              { "type": "string", "maxLength": 100, "minLength": 0 },
              { "type": "null" }
            ],
            "title": "Title",
            "description": "Название задачи",
            "examples": ["Купить продукты"]
          },
          "user_task_id": {
            "type": "integer",
            "title": "User Task Id",
            "description": "ID задачи конкретного пользователя",
            "default": 0,
            "examples": [1]
          }
        },
        "type": "object",
        "required": ["priority"],
        "title": "RequestTask",
        "examples": [
          {
            "completed": false,
            "description": "Купить молоко, хлеб и яйца в магазине",
            "priority": 2,
            "title": "Купить продукты",
            "user_task_id": 1
          }
        ],
        "x-test-info": "Это тестовое поле. Здесь можно добавлять любые свои данные для экспериментов."
      },
      "ResponseTasks": {
        "properties": {
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "description": "Статус выполнения",
            "default": false,
            "examples": [true, false]
          },
          "priority": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": -1.0,
            "title": "Priority",
            "description": "Приоритет задачи (0–3)",
            "examples": [2]
          },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 100, "minLength": 0 },
              { "type": "null" }
            ],
            "title": "Description",
            "description": "Описание задачи",
            "examples": ["Купить молоко, хлеб и яйца в магазине"]
          },
          "title": {
            "anyOf": [
              { "type": "string", "maxLength": 100, "minLength": 0 },
              { "type": "null" }
            ],
            "title": "Title",
            "description": "Название задачи",
            "examples": ["Купить продукты"]
          },
          "user_task_id": {
            "type": "integer",
            "title": "User Task Id",
            "description": "ID задачи конкретного пользователя",
            "default": 0,
            "examples": [1]
          },
          "id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Id",
            "description": "ID задачи глобальный",
            "examples": [1]
          }
        },
        "type": "object",
        "required": ["priority", "id"],
        "title": "ResponseTasks"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/api/auth/token" } }
      }
    }
  }
}